Fixing a NOT NULL constraint
After trying to sign up a user. The system kept on crashing, stating that I am trying to  insert a record into a SQLAlchemy database table called "user," but I'm violating a NOT NULL constraint on the "first_name" & "last_name" column. This means that the "first_name" & "last_name" column cannot be NULL, but I'm attempting to insert a NULL value or not providing any value for it.