Fixed a redirection endpoint error
For some reason flask couldn't build the URL for the endpoint "student.dashboard".

This typically happens when Flask can't find the endpoint you're referring to.

So the solution was that I specify "main.student_dashboard", so to tell Flask to look for the "student_dashboard" endpoint within the main blueprint.

Made this change in my project's  "auth.py" file, and my code worked as expected.